Fix E0308 type mismatch errors in move callbacks
- flowershield.rs: Fix boost call to use slice literal and Option<i32> check
- forestscurse.rs: Fix add_type to use String instead of ID
- frustration.rs: Cast bp to i32 for EventResult::Number
- geomancy.rs: Fix run_event argument wrapping and Option<i32> conversion

+++ src/data/move_callbacks/frustration.rs
@@ -27,6 +27,6 @@ pub fn base_power_callback(battle: &mut Battle, pokemon_pos: (usize, usize), tar
     // || 1 means if bp is 0, return 1
     let bp = if bp == 0 { 1 } else { bp };
 
-    EventResult::Number(bp)
+    EventResult::Number(bp as i32)
 }
